+++ src/Router/Manager/Matcher.php
@@ -95,7 +95,7 @@ trait Matcher
             uksort($result, function ($a, $b) {
                 $a = explode('-', $a);
                 $b = explode('-', $b);
-                return $a[0] !== $b[0] ? $a[0] > $b[0] : $a[1] > $b[1];
+                return $a === $b ? 0 : ($a[0] !== $b[0] ? ($a[0] > $b[0] ? 1 : -1) : ($a[1] > $b[1] ? 1 : -1));
             });
 
             $result = array_values($result);