+++ CHANGELOG.md
@@ -5,7 +5,7 @@ CHANGELOG
 - Use [VSR\\Extend\\Caller](https://github.com/vitorsreis/extend-caller)
 - Cache improvements
   - Add permission control for cache prefixes
-    - flags: FLAG_ROUTER, FLAG_MATCH, FLAG_EXECUTE, FLAG_OTHERS
+    - flags: FLAG_ROUTER, FLAG_MATCH, FLAG_EXECUTE, FLAG_OTHERS, FLAG_ALL
     - allowCache(flag)
     - disallowCache(flag)
 - Fix PSR-12